Fixed the origin of the mask
Found that when I casted the resampled mask array to a float, the origin information was wiped. So I had to respecify the origin once the array had been cast to a float.

Also removed the loop that counts how many '1's there are in the mask array. This may be useful in the future as a way to determine the volume of the tumour (since, after resampling, the voxel volume is 1mm^3, I think).