Create BinaryClassifier.py
Single python script for the Binary Classifier. This was made after it was suggested by our supervisor that Jupyter deals with variables in a strange way. It was proposed that running this script from colab, not from a Jupyter file would mean that our network may be able to run faster.

Co-Authored-By: PHastings37 <63857972+PHastings37@users.noreply.github.com>